Add Scala 3 to the build and check source compatibility (#2456)
* Update publish script

* Remove Scala 3 from publish-impl script

* Add Scala 3 to the build

* Add Scala 3 to the publish script

* Move compiler plugin files to scala-2 subdirectory

* Allow to compile nativelib in Scala 3

* Allow to compile posixlib with Scala 3

* Move Scala 2 overrides to overrides-2 directory

* Revert spurious changes

* Add resolving of scala3lib to sbt plugin

* Update CI workflows, add Scala experimental job

+++ nativelib/src/main/scala/scala/scalanative/runtime/SymbolFormatter.scala
@@ -39,7 +39,7 @@ object SymbolFormatter {
     def readGlobal(): Boolean = read() match {
       case 'M' =>
         readIdent()
-        if (strlen(ident) == 0) {
+        if (strlen(ident) == 0.toUInt) {
           false
         } else {
           strcpy(classNameOut, ident)
@@ -55,7 +55,7 @@ object SymbolFormatter {
         true
       case 'D' | 'P' | 'C' | 'G' =>
         readIdent()
-        if (strlen(ident) == 0) {
+        if (strlen(ident) == 0.toUInt) {
           false
         } else {
           strcpy(methodNameOut, ident)
